+++ src/components/Bubble/utils/index.ts
@@ -1,0 +1,1 @@
+export { sizeMap, fontFamilies, fontSizes, applyStyleToRichText } from './richTextUtils';
